Création d'un fichier descente-du-gradient
